% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_xmart_cols}
\alias{has_xmart_cols}
\title{Check data frame for xMart4 columns}
\usage{
has_xmart_cols(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
bool
}
\description{
Tests to see if the given data frame has all the columns required by xMart4.
The test does not take column order into consideration
(i.e., a,b,c and c,a,b will post them)
}
