% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_wims.R
\name{make_azure_rdir}
\alias{make_azure_rdir}
\title{Create Azure cache directory}
\usage{
make_azure_rdir()
}
\value{
Nothing, but creates directory if it does not already exist.
}
\description{
\code{make_azure_rdir()} is a simple wrapper function that creates the Azure
cache directory on a local machine once, to enable caching of WIMS account
authentication. This replicates the \code{onLoad()} behavior of the AzureAuth
package when called using \code{library(AzureAuth)}.
}
