% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmart4-package.R
\docType{package}
\name{xmart4-package}
\alias{xmart4}
\alias{xmart4-package}
\title{xmart4: R Client for the xMart 4 API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides easy access to the xMart 4 API by managing access tokens to
    UAT and PROD servers automatically, with limited setup required by the user. Currently
    allows systematic access to tables within xMart marts, but additional functionality
    can be provided as requested.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caldwellst/xmart4}
  \item Report bugs at \url{https://github.com/caldwellst/xmart4/issues}
}

}
\author{
\strong{Maintainer}: Seth Caldwell \email{caldwellst@gmail.com}

}
\keyword{internal}
