% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{xmart4_api}
\alias{xmart4_api}
\title{Access the xMart4 API}
\usage{
xmart4_api(
  mart,
  table = NULL,
  xmart_server = c("UAT", "PROD"),
  top = NULL,
  query = NULL,
  token = NULL
)
}
\arguments{
\item{mart}{Name of the xMart4 mart}

\item{table}{Name of a table within the mart, otherwise \code{NULL} (the default)}

\item{xmart_server}{Either 'UAT' (the default) or 'PROD'}

\item{top}{Number of rows of a table to return}

\item{query}{A single string fitting the \href{https://www.odata.org/documentation/odata-version-2-0/uri-conventions/}{Odata protocol} that must start with \code{"$filter="}.}

\item{token}{Access token for xMart4 server. If NULL (the default), the package automatically creates and manages access for the user if Azure client ID and secret set up properly. See \code{vignette("token_setup")} for instructions and details.}
}
\value{
A data frame.
}
\description{
Function that provides access to the xMart4 API. Currently functionality is
limited to accessing lists of tables within marts and the data within them,
but additional functionality can be built-in as requested.
}
