% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mart_access.R
\name{xmart4_mart}
\alias{xmart4_mart}
\title{Get overview of a specific mart}
\usage{
xmart4_mart(
  mart,
  xmart_server = c("UAT", "PROD"),
  auth_type = "client",
  token = NULL
)
}
\arguments{
\item{mart}{Name of the xMart4 mart}

\item{xmart_server}{Either 'UAT' (the default) or 'PROD'}

\item{auth_type}{Type of authorization to use for the token authorization.
If "client" (the default), it uses an AzureAD client setup. The old "wims"
method has been discontinued.}

\item{token}{Access token for xMart4 server. If NULL (the default), the package automatically creates and manages access for the user if Azure client ID and secret set up properly. See \code{vignette("token_setup")} for instructions and details.}
}
\value{
Character vector
}
\description{
Get names of all available tables and views in a single mart.
}
