% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat.R
\name{xmart4_token}
\alias{xmart4_token}
\title{Request xMart4 access token}
\usage{
xmart4_token(client_id = NULL, client_secret = NULL, xmart_server = "UAT")
}
\arguments{
\item{client_id}{Azure client ID}

\item{client_secret}{A}

\item{xmart_server}{Either 'UAT' (the default) or 'PROD'}
}
\value{
List with two values, token and time of expiration, which is 60 minutes
from time of generation.
}
\description{
Using pre-configured Azure client ID and secret, xMart4 access tokens can be generated
to access marts your client has permissions for. Detailed instructions
from the xMart4 development team on setting up a remote client to consume
xMart4 API data is available on the
\href{https://xmartcollaboration.slack.com/files/TJF6QTLE4/F019WGZSSD7?origin_team=TJF6QTLE4}{ xMart4 Slack}.
}
\details{
Unless you need to test multiple clients in your R session, you do not need
to manually create tokens, the package will automatically generate and manage
your tokens if you have properly configured your R environment. Your client ID
and client secret should be set in your .Renviron file. More detailed
instructions can be found on the
\href{https://github.com/caldwellst/xmart4}{ xmart4 package GitHub page}.
}
