% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_client.R
\name{xmart4_token_client}
\alias{xmart4_token_client}
\title{Request xMart4 access token}
\usage{
xmart4_token_client(xmart_id, client_id = NULL, client_secret = NULL)
}
\arguments{
\item{client_id}{Azure client ID. Only required if \code{auth_type == 'client'}.}

\item{client_secret}{Azure client secret. Only required if \code{auth_type == 'client'}.}
}
\description{
Request xMart4 access token

Request xMart4 access token
}
