% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pat.R
\name{xmart4_token_time}
\alias{xmart4_token_time}
\title{Get time left on xMart API token}
\usage{
xmart4_token_time(token)
}
\arguments{
\item{token}{Access token for xMart4 server. If NULL (the default), the package automatically creates and manages access for the user if Azure client ID and secret set up properly. See \code{vignette("token_setup")} for instructions and details.}
}
\value{
Time left until token expiration
}
\description{
xMart API tokens are short-term tokens that last for 60 minutes after generation.
This returns time until token expires. You should only need this if you are
manually managing token generation, which is not recommended.
}
