% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_wims.R
\name{xmart4_token_wims}
\alias{xmart4_token_wims}
\title{Request xMart4 access token}
\usage{
xmart4_token_wims(xmart_id, xmart_server, use_cache = TRUE)
}
\arguments{
\item{xmart_id}{Resource and app ID for authentication.}

\item{xmart_server}{Either 'UAT' (the default) or 'PROD'}

\item{use_cache}{Use Azure cache if TRUE (default), otherwise bypass cache.}
}
\description{
Request xMart4 access token

Request xMart4 access token
}
