% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_general.R
\name{xmart4_token}
\alias{xmart4_token}
\title{Request xMart4 access token}
\usage{
xmart4_token(
  auth_type = c("wims", "client"),
  use_cache = TRUE,
  client_id = NULL,
  client_secret = NULL,
  xmart_server = c("UAT", "PROD")
)
}
\arguments{
\item{auth_type}{Type of authorization to use for the token authorization. For
"wims", the default, this uses the WHO WIMS system to authenticate, with
the AzureAuth package on the backend. If "client", it uses an AzureAD client
setup.}

\item{use_cache}{Use Azure cache if TRUE (default), otherwise bypass cache.}

\item{client_id}{Azure client ID. Only required if \code{auth_type == 'client'}.}

\item{client_secret}{Azure client secret. Only required if \code{auth_type == 'client'}.}

\item{xmart_server}{Either 'UAT' (the default) or 'PROD'}
}
\value{
List with two values, token and time of expiration, which is 60 minutes
from time of generation.
}
\description{
xMart4 access tokens can be generated to access marts either through using a
WHO WIMS account or using pre-configured Azure client ID and secret. The package
automatically manages the tokens behind the scenes for you, so unless you need to test
multiple client access in a single R session, you do not need to manually create tokens.
More detailed instructions can be found on the
\href{https://github.com/caldwellst/xmart4}{ xmart4 package GitHub page}.
}
\details{
For more details on the WIMS and client methods of authentication, see their relevant
vignettes:
\itemize{
\item \href{../doc/token_setup.html}{\code{vignette("token_setup", package = "xmart4")}}
}
}
